requirements.txt
# Requisitos para a Interface Rotor MTM
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
ross-rotordynamics>=1.0.0

# Para análises avançadas
matplotlib>=3.7.0
seaborn>=0.12.0

# Para processamento e salvamento
pickle5
datetime
